enh(all): remove -Wall -Wextra warnings.

+++ core/src/misc/filesystem.cc
@@ -147,7 +147,7 @@ bool filesystem::mkpath(std::string const& path) {
     case ENOENT:
       // parent didn't exist, try to create it
       {
-        int pos = path.find_last_of('/');
+        size_t pos = path.find_last_of('/');
         if (pos == std::string::npos)
           return false;
         if (!mkpath(path.substr(0, pos)))
